chore(snip9): adds option for non snip9 order
