+++ src/index.ts
@@ -1,36 +1,19 @@
 import { formattedDecimalToBigInt, spinSDKClient } from "./utils";
 import { authenticateAccount } from "./auth";
-import { Account, OutsideExecutionVersion, Signer } from "starknet";
-import {
-  CHAIN_ID,
-  coreAddress,
-  coreContract,
-  CURRENCIES,
-  executorAddress,
-  executorContract,
-  rollupInvoker,
-  routerAccount,
-  routerSigner,
-  rpcProvider,
-} from "./constants";
+import { Account, Signer } from "starknet";
+import { CURRENCIES, routerAccount, rpcProvider } from "./constants";
 import readline from "node:readline";
 import {
-  castToApiSignature,
   ExecutionReport,
   OrderConstructor,
-  Result,
   TickerSpecification,
-  TraderSignature,
 } from "layerakira-js";
 import { getTakerOrderStatus, subscribeToExecutionReports } from "./wss";
 import { initializeOrderBuilder } from "./builder";
-import { calcOrderHashAndTypedData, signOrder } from "./signer";
-import { buildSwapOrder, queryGasRelated, queryTransactionHash } from "./swap";
-import { Currency } from "./types";
+import { queryTransactionHash } from "./lib";
 import * as SDK from "layerakira-js";
 import { bigIntReplacer } from "layerakira-js/dist/api/http/utils";
-import { ERC20Token } from "layerakira-js/src/request_types";
-import { SorTaker } from "./sor";
+import { TakerCli } from "./taker";
 import {
   fetchTickerSnapshots,
   SnapshotPathProvider,
@@ -46,7 +29,7 @@ export const routerSDK = spinSDKClient((arg: any) => console.log(arg));
 
 let orderBuilder: OrderConstructor;
 let tickerSpec: Array<TickerSpecification>;
-let sorTaker: SorTaker;
+let taker: TakerCli;
 export let snapshotProvider: SnapshotPathProvider<SDK.Snapshot>;
 
 const setupRouterSDK = async (
@@ -70,7 +53,7 @@ const setupRouterSDK = async (
   orderBuilder = await initializeOrderBuilder(routerSDK);
 };
 
-const setUpSorTaker = async () => {
+const setUpTaker = async () => {
   // making sure that put necessary tokens in the snapshot path provider for sor snapshots
   const tokens = new Set<string>();
   for (const ticker of tickerSpec) {
@@ -82,7 +65,7 @@ const setUpSorTaker = async () => {
     4,
   );
   await fetchTickerSnapshots(routerSDK, tickerSpec);
-  sorTaker = new SorTaker(orderBuilder, console.log);
+  taker = new TakerCli(orderBuilder, console.log);
 };
 
 setupRouterSDK((e) =>
@@ -90,148 +73,8 @@ setupRouterSDK((e) =>
     `Parsed evt: ${getTakerOrderStatus(e)}\n : ${JSON.stringify(e, bigIntReplacer)}`,
   ),
 ).then(() => {
-  setUpSorTaker();
+  setUpTaker();
 });
-
-const swapPipeline = async (
-  payCurrency: Currency,
-  receiveCurrency: Currency,
-  gasTokenCurrency: Currency,
-  lastUpdatedField: "pay" | "receive",
-  payAmount: bigint,
-  receiveAmount: bigint,
-  slippageBips: number = 100,
-  snip9: boolean = true,
-  feeBipsTotal: number = 10_00,
-  useNativeRouter = true,
-): Promise<string> => {
-  let [gasPrice, gasConversionRate] = await queryGasRelated(
-    routerSDK,
-    gasTokenCurrency.code,
-  );
-
-  //Let's assume client call from your frontend "build" method
-  //below part of build method
-  const [order, spendSlippaged] = await buildSwapOrder(
-    clientAccount.address,
-    clientNonce,
-    routerSDK,
-    orderBuilder,
-    tickerSpec,
-    payCurrency,
-    receiveCurrency,
-    gasTokenCurrency,
-    gasPrice,
-    gasConversionRate,
-    lastUpdatedField,
-    payAmount,
-    receiveAmount,
-    slippageBips,
-    snip9,
-    feeBipsTotal,
-  );
-  const [orderHash, orderTypedData] = calcOrderHashAndTypedData(
-    clientAccount!,
-    order,
-    `${CHAIN_ID}`,
-  );
-  let routerSignatureResult: Result<[string, string]>;
-
-  if (useNativeRouter) {
-    // use layerakira router
-    routerSignatureResult = await routerSDK?.akiraHttp.signExternalOrder(order);
-  } else {
-    // use your own router; dont forget to perform necessary preparations
-    routerSignatureResult = {
-      result: castToApiSignature(await routerSigner.signRaw(orderHash)) as [
-        string,
-        string,
-      ],
-    };
-  }
-  // at this stage we properly built order
-  // assume it is end of "build"  endpoint
-
-  if (snip9) {
-    // here we are building snip9 order
-    // a bit more just to avoid any problem with rounding
-    const requiredApproveToSpend = (105n * spendSlippaged) / 100n;
-    const gasFee = order.fee.gas_fee;
-    // make some premium
-    const totalGas =
-      BigInt(order.constraints.number_of_swaps_allowed) *
-      SDK.getGasFeeAndCoin(gasFee, (gasPrice * 105n) / 100n, "STRK", 250)[0];
-
-    const approvals: [[ERC20Token, bigint, string]] = [
-      [payCurrency.code as string, requiredApproveToSpend, executorAddress],
-    ];
-    if (
-      gasTokenCurrency.code != payCurrency.code &&
-      gasTokenCurrency.code != receiveCurrency.code
-    ) {
-      approvals.push([gasFee.fee_token, totalGas, executorAddress]);
-    }
-    // check if client already approved executor
-    const is_approved_executor = await coreContract.call(
-      "is_approved_executor",
-      [clientAccount.address.toString()],
-    );
-    // dont forget in real code to check snip9 version that supported by wallet of user
-    const execOutsidePrimitives = SDK.buildExecuteOutsidePrimitives(
-      executorContract,
-      order,
-      Object.fromEntries(
-        Object.entries(CURRENCIES).map(([key, currency]) => [
-          key,
-          currency.address,
-        ]),
-      ),
-      routerSignatureResult.result!,
-      approvals,
-      rollupInvoker,
-      undefined,
-      undefined,
-      //@ts-ignore
-      !is_approved_executor,
-    );
-    // normally you need check what snip0 revision is supported
-    order.snip9_call = await SDK.buildOutsideExecuteTransaction(
-      //@ts-ignore
-      clientAccount!,
-      execOutsidePrimitives[0],
-      execOutsidePrimitives[1],
-      execOutsidePrimitives[2],
-      OutsideExecutionVersion.V2,
-    );
-  }
-  // at this stage we properly built order
-  // assume it is end of "build" endpoint
-
-  // this is client phase, on frontend where he signs and submits to either layerakira or your relayer backend
-  let sign: TraderSignature = ["0x0", "0x0"];
-  let result: SDK.Result<string>;
-  if (!snip9) {
-    sign = await signOrder(clientAccount!, orderTypedData);
-    result = await routerSDK!.akiraHttp.placeOrder(
-      order,
-      sign,
-      routerSignatureResult.result,
-    );
-  } else {
-    result = await routerSDK!.akiraHttp.placeOrder(
-      order,
-      sign,
-      routerSignatureResult.result,
-    );
-  }
-  // if order passes basic checks -> http would return hash of the order
-  // otherwise error would be returned
-  if (result.result == undefined) {
-    console.warn(`Failed to submit order due ${JSON.stringify(result)}`);
-    return "";
-  }
-  return result.result!;
-};
 
 let clientAccount: Account;
 let clientNonce = 0;
@@ -247,63 +90,18 @@ rl.on("line", (line: string) => {
 
   // List all the avaiable sor paths for this cli
   if (line.startsWith("list_sor_path")) {
-    sorTaker.listSorPaths();
+    taker.listSorPaths();
   }
 
-  // samples:
-  // place_sor STRK 1 AETH 0.147 10 STRK true true
-  // place_sor STRK 1 AUSDT 0.147 10 STRK true true
-  // place_sor AAAVE 1 AETH 0.147 10 AETH true true
-  if (line.startsWith("place_sor")) {
-    let [
-      _,
-      payCode,
-      payAmount,
-      receiveCode,
-      receiveAmount,
-      slippageBips,
-      gasTokenCode,
-      useNativeRouter,
-      externalFunds,
-    ] = line.split(" ");
-    const [payCurrency, receiveCurrency, gasCurrency] = [
-      CURRENCIES[payCode],
-      CURRENCIES[receiveCode],
-      CURRENCIES[gasTokenCode],
-    ];
-    let tweakedField: "pay" | "receive" =
-      formattedDecimalToBigInt(payAmount, payCurrency.unit) > 0n
-        ? "pay"
-        : "receive";
-    const nativeRouter = useNativeRouter.trim().toLowerCase() == "true";
-    sorTaker
-      .placeSorOrder(
-        routerSDK,
-        payCurrency,
-        receiveCurrency,
-        gasCurrency,
-        payAmount,
-        receiveAmount,
-        tweakedField,
-        tickerSpec,
-        clientAccount,
-        clientNonce,
-        externalFunds.trim().toLowerCase() == "true",
-        nativeRouter,
-        parseInt(slippageBips),
-      )
-      .then((hash) => {
-        console.log(`Order hash: ${hash}`);
-      });
-  }
-
-  // samples:
-  // swap AETH 0.1 AUSDC 0 10 AETH true true
-  // swap AUSDC 50 AUSDC 0 10 AETH true true
-  // swap AUSDC 0 AETH 0.01 10 AETH true true
-  // swap AUSDC 5 AETH 0 10 AUSDC true true
-  // fee token in this example always set to receive currency
-  // <payField> <payAmount> <receiveField> <receiveAmount> <slippageBips> <gasToken> <enablesnip9> <useNativeRouter>
+  // samples
+  // swap STRK 1 AETH 0.147 10 STRK true true true
+  // swap STRK 1 AUSDT 0.147 100 STRK true true true
+  // swap AAAVE 1 AETH 0.147 10 AETH true true true
+  // swap AUSDC 50 AUSDT 50 50 AUSDT true true true
+  // swap AETH 0.01 AUSDC 24.15 100 AETH false true true
+  // swap AUSDC 5 AETH 0 10 AUSDC true true true
+  // swap AETH 0.01 AETH 0 0 AETH true true true
+  // <payField> <payAmount> <receiveField> <receiveAmount> <slippageBips> <gasToken> <snip9> <useNativeRouter> <externalFunds>
   if (line.startsWith("swap")) {
     let [
       _,
@@ -315,33 +113,43 @@ rl.on("line", (line: string) => {
       gasTokenCode,
       snip9,
       useNativeRouter,
+      externalFunds,
     ] = line.split(" ");
     const [payCurrency, receiveCurrency, gasCurrency] = [
       CURRENCIES[payCode],
       CURRENCIES[receiveCode],
       CURRENCIES[gasTokenCode],
     ];
-    console.log(gasCurrency, gasTokenCode);
+    if (payCurrency.code === receiveCurrency.code) {
+      console.log("Pay and receive currencies are the same");
+      return;
+    }
+    let tweakedField: "pay" | "receive" =
+      formattedDecimalToBigInt(payAmount, payCurrency.unit) > 0n
+        ? "pay"
+        : "receive";
     const isSnip9 = snip9.trim().toLowerCase() == "true";
     const nativeRouter = useNativeRouter.trim().toLowerCase() == "true";
-    const payAmountRaw = formattedDecimalToBigInt(payAmount, payCurrency.unit);
-    const receiveAmountRaw = formattedDecimalToBigInt(
-      receiveAmount,
-      receiveCurrency.unit,
-    );
-    let tweakedField: "pay" | "receive" = payAmountRaw > 0n ? "pay" : "receive";
-    swapPipeline(
-      payCurrency,
-      receiveCurrency,
-      gasCurrency,
-      tweakedField,
-      payAmountRaw,
-      receiveAmountRaw,
-      parseInt(slippageBips),
-      isSnip9,
-      10_00,
-      nativeRouter,
-    ).then((orderHash) => console.log(`Order hash: ${orderHash}`));
+    taker
+      .placeTakerOrder(
+        routerSDK,
+        payCurrency,
+        receiveCurrency,
+        gasCurrency,
+        payAmount,
+        receiveAmount,
+        tweakedField,
+        tickerSpec,
+        clientAccount,
+        clientNonce,
+        isSnip9,
+        externalFunds.trim().toLowerCase() == "true",
+        nativeRouter,
+        parseInt(slippageBips),
+      )
+      .then((hash) => {
+        console.log(`Order hash: ${hash}`);
+      });
   }
 
   // allows one to query order by its order hash
